backend/requirements.txt
Flask
Flask_SQLAlchemy
Flask_Cors
python-chess
